health_buddy_agent/utils/tools.py
from langchain_core.tools import tool


def icd10_search_tool(query: str) -> str:
    """
    Search for ICD-10 codes based on a query string.
    Returns a string with the search results.
    """
    # Placeholder for actual ICD-10 search logic
    # In a real implementation, this would query an ICD-10 database or API
    return f"ICD-10 codes for '{query}': [A00, A01, A02]"

def recommendation_tool(symptoms: str) -> str:
    """
    Generate recommendations based on the provided symptoms.
    Returns a string with the recommendations.
    """
    # Placeholder for actual recommendation logic
    # In a real implementation, this would analyze symptoms and provide tailored advice
    return f"Recommendations for '{symptoms}': [Rest, Hydration, Over-the-counter medication]"

def urgency_scorer(symptoms: str) -> float:
    """
    Score the urgency of the symptoms on a scale from 0.0 to 1.0.
    Returns a float representing the urgency score.
    """
    # Placeholder for actual urgency scoring logic
    # In a real implementation, this would analyze symptoms and assign an urgency score
    return 0.5