pyproject.toml
[project]
name = "healthbuddyagentapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "langchain[google-genai]>=0.3.26",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "langsmith>=0.4.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
